Change parameter behavior for Clear from preserve to deallocate, which semantically changes the meaning.
Previously the adapter Clear method by default deallocate the object
inside. This is might be not a good default condition, as Adapter might
usually only hold reference to such object.

By calling Clear(true), this will make the Adapter automatically perform
deallocation for every object stored inside it. While calling Clear() or
Clear(false) will leave the object intact.

By default the destructor of Adapter calls Clear(false) and preserve the
object. If the program needs to deallocate Adapter is destroyed. The
program must ensure that every pointer stored inside the adapter is able
to be deallocated or ItemClass implements no deallocator for pointers
which is non-deallocatable.