Addition new Pager component and bugfix in SimpleWebView
1. New Pager component to display simple pager navigation component. It
has top and bottom pager that can be styled using EDC (Currently the
sample for style still in Detikcom project, will be moved here later)

2. Addition on ELM infrastructure to support event handling from signal
received from Edje objects

3. Fix in SimpleWebView which Elm_Box is stupidly enough cannot clear
itself, so it cannot be cleared when we attach new HTML data. So the
workaround is the box is wrapped on another box! Yay!