Introduced SharedEvent which is an automatic allocated event, suitable for asynchronous functions.
Async infrastructure now requires to use this SharedEvent instead of
standard Event, as this will ensure that the Event object is anchored
until the last holding object is destroyed. The event handler attached
to the event must be removed upon the destrcution of the object, so the
event will not call an invalid event handler.

Additional experimental Edje signaling mechanism in GenericList. It
would not working as expected due to the beavior of Elementary Genlist