Add support for URL parameter for REST service
As requested in issue #3

Now the user of REST service can declare parameter for embedded in url
path. The use is the same with the existing parameter declaration, with
addition of ParameterType::URL in the enumeration.

The URL now accept parameter formatting for this purpose. The format is:
http://tempuri.org/{PARAMETER_1}/{PARAMETER_2} with the name within the
bracket is the name registered for parameter declared in the REST
service constructor.