Improvement in tfc_async mechanism
- Now the lambda closure of tfc_async and tfc_async_complete is copied
once (during the calling of tfc_async in user codes) and moved once
(during the serialization of closure to background thread). This removes
unnecessary copy that might happen internally. The lambda itself must be
movable, or the move operation will be replaced by copy automatically by
compiler.
- Add ManagedClass behavior class to be able to track if the class is
still exist or not. The behavior is similar with std::weak_ptr but it is
more lightweight and does not have to create std::shared_ptr to utilize
this. The class itself need not to be managed by any automatic pointer.