[FIX] Invalid Storing rvalue reference of Catch Handler on tfc_async
Although it seems working in no-optimization mode, this is definitely unsafe as the rvalue objects may be optimized out by the compiler which results in dangling reference.

The catch list is now stored directly in concrete value instead of in rvalue reference form to ensure the correctness of building the catch handler objects at the end of evaluating tfc_async expression.