[FIX] Invalid parameter for reflection-based throwing
For propagated exception through GDBusClient, the Throw function call of reflected type info is passed an l-value reference to string object, which makes the correct constructor for exception could not be found as there is no l-value string constructor.

It is now changed by passing char const* parameter so it will invoke the C-string style constructor.